Check the minimum amount before defining the fromAmount
The min_amount returned from LetsExchange is the same currency that the user requested the amount as. For example, if the user requested to sell ETH to receive 5 BNB (and pretending this didn't meet the minimum) then the API will return the min_amount as BNB. We need to check the min_amount within the same context of either reply from the API before moving on.