DEXs: Preserve customNetworkFee from first order for max send order
The DEXs calculate the gas price themselves and set a custom fee. To avoid the second order (the max amount) we can hold onto the first gas price to avoid the chance that the gas price goes up during the second order and therefore invalidating the transaction. It's worth pointing out that actually spending the max amount of the native currency isn't actually possible and will always leave some dust due to the gas limit overestimation but this change means we won't accidentally try to spend beyond our balance.

The bug exists across the board but is only currently solvable for swap plugins that set fees themselves.