Godex: Flip minimum amount and network checks
The response for requested swaps below the minimums does not include chain info so it was getting caught in the unsupported error flow. I pinged Godex for clarity on how to tell which networks the min_amount applied to since we don't get to specify that before receiving the minimum error.

As far as I can tell, it is possible the returned min_amount might be for a different network than the user requests but it's probably better to show that than completely unsupported.