Use the value derived from the smart contract method instead of the populated Ethers transaction
The value field cannot be relied upon for token transactions since, generally, this field is used for mainnet ETH amount. Instead, we should use the amount we pass to Ethers as the nativeAmount. We also need to account for the fact that in the case where there are two transactions the first one is a 0 amount approval tx and only the second one needs the amount.