Fix LetsExchange special mainnet handling
LetsExchange uses multiple network names for some chains. For example, ETH is the valid network name when trading ETH on Ethereum but if trading for a token then the network is called ERC20. We can't make any assumptions about the code because other token-bearing chains like ALGO always use the network code ALGO.