Remove networkFee for a 0xgasless swap
Fee is paid out of the fromToken amount, so in other words, the exchange
rate is adjusted to account for the fee. This means the user will always
expect to pay the `nativeAmount` with no fee, but this does impact the
exchange rate used to calculate the buy-token (to-token) amount.