Add tokenId mapping to Changenow
"Draw the rest of the fucking owl"

The goal is to map the swap provider's chain/token code combo to EdgeTokenIds without causing too much of a slow down by adding unnecessary fetches. To accomplish this, we can grab the data from the provider once and hold them in memory indefinitely as `chainCodeTickerMap`. We'll convert the cached data As requests are made, we map that cached data to EdgeTokenIds using `getTokenId` which is available on the wallets in the request.