Add additional approval transaction for Ethereum USDT
USDT requires an allowance to be zero before it can be nonzero. Most, like 99% of tokens, don't require this behavior.

We would ideally check the allowance first before adding any approval transactions to the swap to avoid unnecessary transactions. Instead of adding a bunch of node lists and queries here maybe we could plumb something through the engine or tools to access those node lists.